source/pyproject.toml
[tool.poetry]
name = "progress"
version = "0.1.0"
description = ""
authors = ["SagidullinEmil <se12373@auca.kg>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
pillow = "^10.2.0"
django-jet-reboot = "^1.3.7"
google-api-python-client = "1.4.1"
django-ckeditor-5 = "^0.2.11"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
